+++ src/store/store.ts
@@ -19,7 +19,7 @@ import layerDetailsModal from './layerDetailsModal';
 import layerTree from './layerTree';
 import legal from './legal';
 import logoPath from './logoPath';
-import mapToolbarVisible from './mapToolbarVisible';
+import mapToolbar from './mapToolbar';
 import newsTextIds from './newsText';
 import print from './print';
 import searchEngines from './searchEngines';
@@ -31,6 +31,7 @@ import toolMenu from './toolMenu';
 import uploadDataModal from './uploadDataModal';
 import user from './user';
 import userMenu from './userMenu';
+import measure from './measure';
 
 export type AsyncReducer = Record<string, Reducer>;
 
@@ -50,7 +51,8 @@ export const createReducer = (asyncReducers?: AsyncReducer) => {
     layerTree,
     legal,
     logoPath,
-    mapToolbarVisible,
+    mapToolbar,
+    measure,
     newsTextIds,
     print,
     searchEngines,